Fixed layout in the component SelectedProductFilter
